requirements.txt
flask==2.0.1
selenium==4.15.2
gunicorn==20.1.0
webdriver-manager==4.0.1
requests==2.31.0
google-cloud-aiplatform==1.25.0
chromedriver-autoinstaller==0.6.2